RUM-2708: Rename packages from base64 to resources
